[Fix] 대시보드 정보 api 요청 위치 수정, 라우트 경로 수정
