+++ src/app/api/dashboards/[dashboardId]/route.ts
@@ -10,10 +10,6 @@ export const GET = async (req: Request, { params }: { params: IParams }) => {
   const { dashboardId } = params;
   const id = Number(dashboardId);
 
-  const { searchParams } = new URL(req.url);
-  const page = Number(searchParams.get("page")) || 1;
-  const size = Number(searchParams.get("size")) || 10;
-
   const cookieStore = cookies();
   const token = cookieStore.get("accessToken")?.value;
 
@@ -26,14 +22,15 @@ export const GET = async (req: Request, { params }: { params: IParams }) => {
   }
 
   try {
-    const response = await apiClient.get(`/columns?dashboardId=${id}&page=${page}&size=${size}`, {
+    //https://sp-taskify-api.vercel.app/9-4/dashboards/12067
+    const response = await apiClient.get(`/dashboards/${id}`, {
       headers: {
         Authorization: `Bearer ${token}`,
       },
     });
 
     if (response.status === 200) {
-      const data = response.data.data;
+      const data = response.data;
       return NextResponse.json(data, { status: 200 });
     }
   } catch (error) {
@@ -43,6 +40,7 @@ export const GET = async (req: Request, { params }: { params: IParams }) => {
     }
   }
 };
+
 // 대시보드 수정 api
 export const PUT = async (req: Request) => {
   const { searchParams } = new URL(req.url);
